+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 * Can see the unlocking amount transaction when an order is matched on the price less than was locked
 * 500 on get history (create manage limits request)
 * participants statistics for immediate sale
+* empty sale participants effects in admin panel
+* participant effects for manage offer and payment requests
 
 ## 3.7.1